+++ pom.xml
@@ -16,7 +16,9 @@
 	<properties>
 		<java.version>17</java.version>
 	</properties>
+	
 	<dependencies>
+		
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
@@ -25,7 +27,6 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
-
 		<dependency>
       		<groupId>com.mysql</groupId>
       		<artifactId>mysql-connector-j</artifactId>
@@ -41,6 +42,12 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+		    <groupId>org.springframework.boot</groupId>
+		    <artifactId>spring-boot-starter-thymeleaf</artifactId>
+		    <version>3.2.1</version>
+		</dependency>		
+		
 	</dependencies>
 
 	<build>
